FRONTEND2/backend/requirements.txt
flask==2.2.5
flask-cors==3.0.10

pandas==2.1.2
scikit-learn==1.3.2
joblib==1.3.2
opencv-python-headless==4.8.0.74
Pillow==10.1.0
fer==22.4.0
mtcnn==0.1.1